refactor: signup test same as login test
- remove wrapper methods
- inlined invalid data
- moved the DTO files one level up
